pip3-requirements.txt
alabaster==0.7.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
awscliv2==2.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.28
botocore==1.27.59
bottle==0.12.23
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
clang-format==14.0.6
click==8.1.3
cloudpickle==2.2.0
cmake-format==0.6.11
colorama==0.4.6
coloredlogs==15.0.1
conan==1.53.0
cryptography==38.0.1
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
docutils==0.19
e==1.4.5
entrypoints==0.4
executing==0.8.3
executor==23.2
fastapi==0.87.0
fasteners==0.18
fastjsonschema==2.16.1
flake8==5.0.4
fonttools==4.34.4
greenlet==1.1.1
gym==0.26.2
gym-notices==0.0.8
humanfriendly==10.0
ibapi==9.81.1.post1
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
json5==0.9.6
jsonschema==4.17.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
node-semver==0.6.1
notebook==6.4.12
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.23.4
packaging==21.3
pandas==1.4.3
pandoc==2.2
pandocfilters==1.5.0
parso==0.8.3
patch-ng==1.17.4
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
pluginbase==1.0.1
plumbum==1.8.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
property-manager==3.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.12.0
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
s3transfer==0.6.0
scipy==1.9.3
seaborn==0.12.0
semantic-version==2.10.0
Send2Trash==1.8.0
setuptools==65.5.0
setuptools-rust==1.3.0
sip==6.6.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.39
stack-data==0.3.0
starlette==0.21.0
terminado==0.15.0
testresources==2.0.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
verboselogs==1.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==3.6.1
zipp==3.10.0
